handle rekog response and continue building out and fixing shared_helpers 2. add helper script to delete lambda layer versions
