+++ shared_helpers/2-package.sh
@@ -5,4 +5,5 @@ mkdir -p python
 cp -r create_layer/lib python/
 zip -r layer_content.zip python
 
+rm -f ../serverless/layer/shared/layer_content.zip
 cp layer_content.zip ../serverless/layer/shared/
